# Remove useless {S}