@odow 2021-05-05
TIL:
```Julia
julia> foo(x) where {T} = x^2
foo (generic function with 1 method)

julia> bar(x) = x^2
bar (generic function with 1 method)

julia> @code_native foo(1)
	.section	__TEXT,__text,regular,pure_instructions
; ┌ @ REPL[1]:1 within `foo'
	pushq	%rax
	movq	%rsi, (%rsp)
	movq	(%rsi), %rax
; │┌ @ intfuncs.jl:312 within `literal_pow'
; ││┌ @ int.jl:88 within `*'
	movq	(%rax), %rdi
	imulq	%rdi, %rdi
; │└└
	movabsq	$jl_box_int64, %rax
	callq	*%rax
	popq	%rcx
	retq
; └

julia> @code_native bar(1)
	.section	__TEXT,__text,regular,pure_instructions
; ┌ @ REPL[2]:1 within `bar'
	movq	%rdi, %rax
; │┌ @ intfuncs.jl:312 within `literal_pow'
; ││┌ @ int.jl:88 within `*'
	imulq	%rdi, %rax
; │└└
	retq
	nopl	(%rax,%rax)
; └
```